// packages/sales-pricing/src/schemas/schema.types.ts
export type Maybe<T> = T | null;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  Decimal: { input: any; output: any; }
  JSON: { input: any; output: any; }
  NonEmptyString: { input: any; output: any; }
  NonNegativeDecimal: { input: any; output: any; }
  NonNegativeInt: { input: any; output: any; }
  PositiveDecimal: { input: any; output: any; }
  PositiveInt: { input: any; output: any; }
  _Any: { input: any; output: any; }
  _FieldSet: { input: any; output: any; }
};

export type BaseOptionsInput = {
  limit?: InputMaybe<Scalars['NonNegativeInt']['input']>;
  offset?: InputMaybe<Scalars['NonNegativeInt']['input']>;
  prefix?: InputMaybe<Scalars['NonEmptyString']['input']>;
};

export type BasicPricingObject = {
  __typename?: 'BasicPricingObject';
  id: Scalars['PositiveInt']['output'];
  name: Scalars['NonEmptyString']['output'];
};

export type BasicPricingProduct = {
  __typename?: 'BasicPricingProduct';
  area: Scalars['NonNegativeDecimal']['output'];
  id: Scalars['PositiveInt']['output'];
  number: Scalars['NonEmptyString']['output'];
  one_gt_id?: Maybe<Scalars['PositiveInt']['output']>;
  price: Scalars['NonNegativeDecimal']['output'];
};

export type BasicPricingProductCategory = {
  __typename?: 'BasicPricingProductCategory';
  id: Scalars['PositiveInt']['output'];
  name: Scalars['NonEmptyString']['output'];
};

export type BasicPricingProductType = {
  __typename?: 'BasicPricingProductType';
  id: Scalars['PositiveInt']['output'];
  name: Scalars['NonEmptyString']['output'];
};

export type GetPricingProductInput = {
  id: Scalars['PositiveInt']['input'];
};

export type GetPricingProductsInput = {
  object_id: Scalars['PositiveInt']['input'];
  options?: InputMaybe<BaseOptionsInput>;
};

export type PricingProduct = {
  __typename?: 'PricingProduct';
  category: BasicPricingProductCategory;
  object: BasicPricingObject;
  product: BasicPricingProduct;
  product_type?: Maybe<BasicPricingProductType>;
};

export type PricingProducts = {
  __typename?: 'PricingProducts';
  products: Array<PricingProduct>;
  total_count: Scalars['PositiveInt']['output'];
};

export type Query = {
  __typename?: 'Query';
  _service: _Service;
  getPricingProduct: PricingProduct;
  getPricingProducts: PricingProducts;
};


export type QueryGetPricingProductArgs = {
  input: GetPricingProductInput;
};


export type QueryGetPricingProductsArgs = {
  input: GetPricingProductsInput;
};

export type _Service = {
  __typename?: '_Service';
  sdl?: Maybe<Scalars['String']['output']>;
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;



/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  BaseOptionsInput: BaseOptionsInput;
  BasicPricingObject: ResolverTypeWrapper<BasicPricingObject>;
  BasicPricingProduct: ResolverTypeWrapper<BasicPricingProduct>;
  BasicPricingProductCategory: ResolverTypeWrapper<BasicPricingProductCategory>;
  BasicPricingProductType: ResolverTypeWrapper<BasicPricingProductType>;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
  DateTime: ResolverTypeWrapper<Scalars['DateTime']['output']>;
  Decimal: ResolverTypeWrapper<Scalars['Decimal']['output']>;
  GetPricingProductInput: GetPricingProductInput;
  GetPricingProductsInput: GetPricingProductsInput;
  JSON: ResolverTypeWrapper<Scalars['JSON']['output']>;
  NonEmptyString: ResolverTypeWrapper<Scalars['NonEmptyString']['output']>;
  NonNegativeDecimal: ResolverTypeWrapper<Scalars['NonNegativeDecimal']['output']>;
  NonNegativeInt: ResolverTypeWrapper<Scalars['NonNegativeInt']['output']>;
  PositiveDecimal: ResolverTypeWrapper<Scalars['PositiveDecimal']['output']>;
  PositiveInt: ResolverTypeWrapper<Scalars['PositiveInt']['output']>;
  PricingProduct: ResolverTypeWrapper<PricingProduct>;
  PricingProducts: ResolverTypeWrapper<PricingProducts>;
  Query: ResolverTypeWrapper<{}>;
  String: ResolverTypeWrapper<Scalars['String']['output']>;
  _Any: ResolverTypeWrapper<Scalars['_Any']['output']>;
  _FieldSet: ResolverTypeWrapper<Scalars['_FieldSet']['output']>;
  _Service: ResolverTypeWrapper<_Service>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  BaseOptionsInput: BaseOptionsInput;
  BasicPricingObject: BasicPricingObject;
  BasicPricingProduct: BasicPricingProduct;
  BasicPricingProductCategory: BasicPricingProductCategory;
  BasicPricingProductType: BasicPricingProductType;
  Boolean: Scalars['Boolean']['output'];
  DateTime: Scalars['DateTime']['output'];
  Decimal: Scalars['Decimal']['output'];
  GetPricingProductInput: GetPricingProductInput;
  GetPricingProductsInput: GetPricingProductsInput;
  JSON: Scalars['JSON']['output'];
  NonEmptyString: Scalars['NonEmptyString']['output'];
  NonNegativeDecimal: Scalars['NonNegativeDecimal']['output'];
  NonNegativeInt: Scalars['NonNegativeInt']['output'];
  PositiveDecimal: Scalars['PositiveDecimal']['output'];
  PositiveInt: Scalars['PositiveInt']['output'];
  PricingProduct: PricingProduct;
  PricingProducts: PricingProducts;
  Query: {};
  String: Scalars['String']['output'];
  _Any: Scalars['_Any']['output'];
  _FieldSet: Scalars['_FieldSet']['output'];
  _Service: _Service;
};
// ...
export type BasicPricingObjectResolvers<ContextType = any, ParentType extends ResolversParentTypes['BasicPricingObject'] = ResolversParentTypes['BasicPricingObject']> = {
  id?: Resolver<ResolversTypes['PositiveInt'], ParentType, ContextType>;
  name?: Resolver<ResolversTypes['NonEmptyString'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type BasicPricingProductResolvers<ContextType = any, ParentType extends ResolversParentTypes['BasicPricingProduct'] = ResolversParentTypes['BasicPricingProduct']> = {
  area?: Resolver<ResolversTypes['NonNegativeDecimal'], ParentType, ContextType>;
  id?: Resolver<ResolversTypes['PositiveInt'], ParentType, ContextType>;
  number?: Resolver<ResolversTypes['NonEmptyString'], ParentType, ContextType>;
  one_gt_id?: Resolver<Maybe<ResolversTypes['PositiveInt']>, ParentType, ContextType>;
  price?: Resolver<ResolversTypes['NonNegativeDecimal'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type BasicPricingProductCategoryResolvers<ContextType = any, ParentType extends ResolversParentTypes['BasicPricingProductCategory'] = ResolversParentTypes['BasicPricingProductCategory']> = {
  id?: Resolver<ResolversTypes['PositiveInt'], ParentType, ContextType>;
  name?: Resolver<ResolversTypes['NonEmptyString'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type BasicPricingProductTypeResolvers<ContextType = any, ParentType extends ResolversParentTypes['BasicPricingProductType'] = ResolversParentTypes['BasicPricingProductType']> = {
  id?: Resolver<ResolversTypes['PositiveInt'], ParentType, ContextType>;
  name?: Resolver<ResolversTypes['NonEmptyString'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};
// ...
export type PricingProductResolvers<ContextType = any, ParentType extends ResolversParentTypes['PricingProduct'] = ResolversParentTypes['PricingProduct']> = {
  category?: Resolver<ResolversTypes['BasicPricingProductCategory'], ParentType, ContextType>;
  object?: Resolver<ResolversTypes['BasicPricingObject'], ParentType, ContextType>;
  product?: Resolver<ResolversTypes['BasicPricingProduct'], ParentType, ContextType>;
  product_type?: Resolver<Maybe<ResolversTypes['BasicPricingProductType']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type PricingProductsResolvers<ContextType = any, ParentType extends ResolversParentTypes['PricingProducts'] = ResolversParentTypes['PricingProducts']> = {
  products?: Resolver<Array<ResolversTypes['PricingProduct']>, ParentType, ContextType>;
  total_count?: Resolver<ResolversTypes['PositiveInt'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  _service?: Resolver<ResolversTypes['_Service'], ParentType, ContextType>;
  getPricingProduct?: Resolver<ResolversTypes['PricingProduct'], ParentType, ContextType, RequireFields<QueryGetPricingProductArgs, 'input'>>;
  getPricingProducts?: Resolver<ResolversTypes['PricingProducts'], ParentType, ContextType, RequireFields<QueryGetPricingProductsArgs, 'input'>>;
};

export interface _AnyScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['_Any'], any> {
  name: '_Any';
}

export interface _FieldSetScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['_FieldSet'], any> {
  name: '_FieldSet';
}

export type _ServiceResolvers<ContextType = any, ParentType extends ResolversParentTypes['_Service'] = ResolversParentTypes['_Service']> = {
  sdl?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type Resolvers<ContextType = any> = {
  BasicPricingObject?: BasicPricingObjectResolvers<ContextType>;
  BasicPricingProduct?: BasicPricingProductResolvers<ContextType>;
  BasicPricingProductCategory?: BasicPricingProductCategoryResolvers<ContextType>;
  BasicPricingProductType?: BasicPricingProductTypeResolvers<ContextType>;
  DateTime?: GraphQLScalarType;
  Decimal?: GraphQLScalarType;
  JSON?: GraphQLScalarType;
  NonEmptyString?: GraphQLScalarType;
  NonNegativeDecimal?: GraphQLScalarType;
  NonNegativeInt?: GraphQLScalarType;
  PositiveDecimal?: GraphQLScalarType;
  PositiveInt?: GraphQLScalarType;
  PricingProduct?: PricingProductResolvers<ContextType>;
  PricingProducts?: PricingProductsResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
  _Any?: GraphQLScalarType;
  _FieldSet?: GraphQLScalarType;
  _Service?: _ServiceResolvers<ContextType>;
};